02. Recognition/06. based Arcface/requirements.txt
absl-py==0.15.0
asttokens @ file:///home/conda/feedstock_root/build_artifacts/asttokens_1618968359944/work
astunparse==1.6.3
backcall @ file:///home/conda/feedstock_root/build_artifacts/backcall_1592338393461/work
backports.functools-lru-cache @ file:///home/conda/feedstock_root/build_artifacts/backports.functools_lru_cache_1618230623929/work
beautifulsoup4==4.10.0
bs4==0.0.1
cachetools==5.0.0
certifi==2021.10.8
charset-normalizer==2.0.12
cloudpickle==2.0.0
colorama @ file:///home/conda/feedstock_root/build_artifacts/colorama_1602866480661/work
cycler==0.11.0
debugpy @ file:///D:/bld/debugpy_1649586564473/work
decorator @ file:///home/conda/feedstock_root/build_artifacts/decorator_1641555617451/work
dlib @ file:///D:/bld/dlib_1647301421331/work
dm-tree==0.1.7
entrypoints @ file:///home/conda/feedstock_root/build_artifacts/entrypoints_1643888246732/work
executing @ file:///home/conda/feedstock_root/build_artifacts/executing_1646044401614/work
flatbuffers==1.12
fonttools==4.30.0
gast==0.4.0
google-auth==2.6.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.34.1
h5py==3.1.0
idna==3.3
imageio==2.19.2
importlib-metadata==4.11.3
ipykernel @ file:///D:/bld/ipykernel_1649684468721/work/dist/ipykernel-6.13.0-py3-none-any.whl
ipython @ file:///D:/bld/ipython_1648413716235/work
jedi @ file:///D:/bld/jedi_1649067326950/work
joblib==1.1.0
jupyter-client @ file:///home/conda/feedstock_root/build_artifacts/jupyter_client_1649340884673/work
jupyter-core @ file:///D:/bld/jupyter_core_1645024505069/work
keras-nightly==2.5.0.dev2021032900
Keras-Preprocessing==1.1.2
kiwisolver==1.4.0
lxml==4.8.0
Markdown==3.3.6
matplotlib==3.5.1
matplotlib-inline @ file:///home/conda/feedstock_root/build_artifacts/matplotlib-inline_1631080358261/work
natsort==8.1.0
nest-asyncio @ file:///home/conda/feedstock_root/build_artifacts/nest-asyncio_1648959695634/work
networkx==2.8
numpy @ file:///D:/bld/numpy_1646717636399/work
oauthlib==3.2.0
opencv-python==4.2.0.34
opt-einsum==3.3.0
packaging @ file:///home/conda/feedstock_root/build_artifacts/packaging_1637239678211/work
parso @ file:///home/conda/feedstock_root/build_artifacts/parso_1638334955874/work
pickleshare @ file:///home/conda/feedstock_root/build_artifacts/pickleshare_1602536217715/work
Pillow==9.0.1
prompt-toolkit @ file:///home/conda/feedstock_root/build_artifacts/prompt-toolkit_1649130487073/work
protobuf==3.19.4
psutil @ file:///D:/bld/psutil_1648873144269/work
pure-eval @ file:///home/conda/feedstock_root/build_artifacts/pure_eval_1642875951954/work
pyasn1==0.4.8
pyasn1-modules==0.2.8
Pygments @ file:///home/conda/feedstock_root/build_artifacts/pygments_1641580240686/work
pyparsing @ file:///home/conda/feedstock_root/build_artifacts/pyparsing_1649603503565/work
pytesseract==0.3.9
python-dateutil @ file:///home/conda/feedstock_root/build_artifacts/python-dateutil_1626286286081/work
PyWavelets==1.3.0
pywin32==303
PyYAML==6.0
pyzmq @ file:///D:/bld/pyzmq_1649055812949/work
requests==2.27.1
requests-oauthlib==1.3.1
rsa==4.8
scikit-image==0.19.2
scikit-learn==1.0.2
scipy==1.8.0
silence-tensorflow==1.1.1
six @ file:///home/conda/feedstock_root/build_artifacts/six_1620240208055/work
sklearn==0.0
soupsieve==2.3.1
split-folders==0.5.1
stack-data @ file:///home/conda/feedstock_root/build_artifacts/stack_data_1644872665635/work
tensorboard==2.8.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow-estimator==2.5.0
tensorflow-gpu==2.5.0
tensorflow-probability==0.16.0
termcolor==1.1.0
threadpoolctl==3.1.0
tifffile==2022.5.4
tornado @ file:///D:/bld/tornado_1648827448708/work
tqdm==4.63.0
traitlets @ file:///home/conda/feedstock_root/build_artifacts/traitlets_1635260543454/work
typing-extensions==3.7.4.3
urllib3==1.26.8
wcwidth @ file:///home/conda/feedstock_root/build_artifacts/wcwidth_1600965781394/work
Werkzeug==2.0.3
wincertstore==0.2
wrapt==1.12.1
zipp==3.7.0
